Intents are now loaded as classes
Previously they were loaded as objects. Now, they're loaded and stored
as classes by the bot, and instanciated (read: turned into an object)
when they are used. In other words, each time a user uses the "Hello"
intent, a Hello object is created. That way, data can be safely stored
in this object, knowing that no other user will use it, and this user
will not use it again.